Don't import estimate if pytorch is not available
