+++ causing/__init__.py
@@ -4,8 +4,15 @@
 # flake8: noqa
 
 # public Causing API
-from causing.estimate import estimate_models
+from causing.bias import estimate_biases
 from causing.indiv import create_indiv
 from causing.graph import create_json_graphs
 from causing.model import Model
 from causing.simulate import simulate, SimulationParams
+
+try:
+    import torch
+    from causing.estimate import estimate_models
+except:
+    # Running without pytorch is allowed, but estimate_models won't be available
+    pass
