+++ causing/graph.py
@@ -6,8 +6,32 @@ from itertools import chain
 import numpy as np
 from numpy import allclose, isnan
 from causing import utils
-from causing.estimate import tvals
 from causing.model import Model
+
+
+def tvals(eff, std):
+    """compute t-values by element wise division of eff and std matrices"""
+
+    assert eff.shape == std.shape
+
+    if len(eff.shape) == 1:  # vector
+        rows = eff.shape[0]
+        tvalues = np.empty(rows) * np.nan
+        for i in range(rows):
+            if std[i] != 0:
+                tvalues[i] = eff[i] / std[i]
+
+    if len(eff.shape) == 2:  # matrix
+        rows, cols = eff.shape
+        tvalues = np.zeros((rows, cols))
+        for i in range(rows):
+            for j in range(cols):
+                if std[i, j] != 0:
+                    tvalues[i, j] = eff[i, j] / std[i, j]
+                else:
+                    tvalues[i, j] = np.nan
+
+    return tvalues
 
 
 def color_scheme(value, base):
